added Binary search with bubble sort inside

+++ BfsSearcher.h
@@ -6,6 +6,6 @@ namespace silber {
 	class BfsSearcher : public ISearcher {
 	public:
 		BfsSearcher() = default;
-		virtual void search(SearchInfo& info, const Matrix<uint8_t>& data) override;
+		virtual void search(SearchInfo& info, const Matrix<uint16_t>& data) override;
 	};
 }
