+++ DfsSearcher.h
@@ -6,6 +6,6 @@ namespace silber {
 	class DfsSearcher : public ISearcher {
 	public:
 		DfsSearcher() = default;
-		virtual void search(SearchInfo& info, const Matrix<uint8_t>& data) override;
+		virtual void search(SearchInfo& info, const Matrix<uint16_t>& data) override;
 	};
 }
